Create app-not-secure
In this unsecure version

Removed the use of bcrypt library for password hashing.
Removed the hash_password method responsible for generating salt and hashing the password.
Stored the passwords in plain text format in the database.
Removed the check_password method that verifies the password against the stored hashed password.
Again, please note that storing passwords in plain text is highly discouraged in real-world scenarios due to security vulnerabilities. This unsecure version is solely meant for educational purposes to understand the importance of proper password hashing and security practices.